♻️  useTab refactor
