+++ src/components/map/map-body/search-location/search-location-type/SearchLocationType.tsx
@@ -1,5 +1,5 @@
 import { Button } from '@/components/ui/button';
-import useTab from '@/hooks/client/map/search-location/useTab';
+import useSearchLocationType from '@/hooks/client/map/search-location/useSearchLocationType';
 import { cn } from '@/lib/utils';
 
 const CONTENT_TYPE = [
@@ -11,10 +11,13 @@ const CONTENT_TYPE = [
     tab: '산책로',
     CALL_TYPE: 'walk',
   },
-] as const;
-export type T_Search = typeof CONTENT_TYPE;
+];
 const SearchLocationType = () => {
-  const { changeItem, currentIndex } = useTab({ initialValue: 0, allTabs: CONTENT_TYPE });
+  const { changeItem, currentIndex } = useSearchLocationType({
+    initialValue: 0,
+    allTabs: CONTENT_TYPE,
+    type: 'CALL_TYPE',
+  });
   return (
     <div className="border-b-2 border-[#C5C9CF] ">
       {CONTENT_TYPE.map((content, idx) => (
